Setup init-test
